Keep the navigation bar from getting too low or too high.
This simple change keeps the navigation bar from going into the screen
or leaving the screen by too much.

You will notice that I added +20 to the contraints.  When a navigation
controller is a top level view controller (not nested inside another
view or view controller, the navigation bar is offset by the status bar
height.  We could in the future add logic to better account for the
different scenarios.  Think "convertPoint:...".