Make the navigation bar translucent.
This change extends the reach of navigation child view controllers up
to the top of the navigation controller, behind the navigation bar.
For UITableViewController subclasses, the navigation controller
automatically adjusts the content insets to keep the content aligned
correctly.  Other views will need to be adjusted accordingly.

This has the additional effect of making the navigation bar
translucent.  You may not notice it right away, but if you scroll down,
then up, you will see the table cells behind the navigation bar.  If
this is undesirable, a custom background can be applied.