Return the navigation bar to the initial position on navigation actions.
The new MagicScrollingNavigationController hooks into the navigation
events and instantly moves the navigation bar to the default position.

Because we are using a translucent header, the table view's content
insets are set on navigation triggering a scroll event and adjusting
the navigation bar location.  We overcome this by attaching adding
logic to only listen to scroll events if they happen between  the start
of drag and the end of drag or deceleration.

We can make it prettier by animating the position change.