Animate the navigation bar in on navigation.
Using slow animations, we can adjust the animation duration to match
the navigation.  This ends up being 0.35 seconds.

Because animations can be fickle, we do the minimum amount possible.
If the position will change a percievable amount and we're animating,
animate a transition to the new position, otherwise jump there.