Refactor:  Move logic about navigation bar placement into the navigation controller.
Now that the navigation bar is working, there are a number of things we
may want to do.  We may want to have the navigation bar appear on tap,
like the New York Times app.  We may want the navigation bar to appear
when scroll completes, like Fab.  Just as polish, we will want to make
it so the navigation bar isn't just partially visible.

Most importantly, we don't want to reimplement this for every
controller.  To make things simple, we will move the scroll logic to
the navigation controller.  Children can either call the methods or
make the navigation controller the delegate of their scroll view.