Prevent the navigation bar from being partially on screen.
This is one of our first opportunities to polish the experience and its
completely up to you what we do.  We will just add some simple logic to
keep the navigation bar from being partially obscured when scrolling
stops.  To make things simple, we will just check if the navigation bar
is onscreen and if so show the bar at the default speed.