Testing it out:  Does it play nicely with other features?  Yes it does.
At some point, you may want to hide the navigation controller,
introducing the question: Does it play nicely with other features?

The answer is Yes!  It works very well with the main property you would
apply to the navigation bar, navigationBarHidden.  Throughout our code,
we changed only one value in the whole navigation controller, the
y-value of the navigation bar's center.  This means the only conflict
is in internal logic around this value.  Fortunately, there doesn't
appear to be much and the navigation bar happily slides into the right
position when you show or hide it.

It is also a cinch to remove completely, simply change the app delegate
to use a vanilla UINavigationController and see how your app still
works as expected, just with the navigation bar fixed to the top.