Enhancement:  Always reveal the navigation barwhen scrolling up.
Up to now, we focused simply on the behavior while scrolling down
(navigation bar is hidden) or scrolling up (navigation bar is shown.
What about when you do both, you scroll down then up?  If you try it on
the previous build, the navigation bar will not be shown until you
scroll back to the original position.  To get around this, we will need
to thing of the navigation bar change as change since last scroll event
rather thant change since start of scrolling.  This is pretty easy to
do and simplifies the code.